requirements.txt
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
numpy==1.25.0
polyline==2.0.0
protobuf==3.20.3
pydantic==1.10.11
python-dotenv==1.0.0
requests==2.31.0
shapely==2.0.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
